backend/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
Flask-CORS
Flask-JWT-Extended
Werkzeug
requests
pypdf
reportlab
google-generativeai 